% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{esteps}
\alias{esteps}
\title{esteps}
\usage{
esteps(s, m, bt, st, d)
}
\arguments{
\item{s}{The number of steps per revolution of your motor.}

\item{m}{Your driver's microstep setting.}

\item{bt}{The number of teeth on your big gear.}

\item{st}{The number of teeth onyour small gear.}

\item{d}{The diameter of your hobbed bolt.}
}
\description{
Calculate proper steps for your E motor in steps/mm.
}
\examples{
esteps(200, 16, 39, 11, 7)
}

