\name{printtools3d-package}
\alias{printtools3d-package}
\alias{printtools3d}
\docType{package}
\title{
\packageTitle{printtools3d}
}
\description{
\packageDescription{printtools3d}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{printtools3d}
\packageIndices{printtools3d}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{printtools3d}

Maintainer: \packageMaintainer{printtools3d}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
