% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resteps.R
\name{resteps}
\alias{resteps}
\title{printtools3d}
\usage{
resteps(o, d)
}
\arguments{
\item{o}{The original setting for your E steps.}

\item{d}{The diameter of your hobbed bolt.}
}
\description{
Refine your steps for your E motor in steps/mm.  Mark your filament 150mm above your extruder, heat up your extruder, then extrude 100mm of filament with the GCODE G1 E100.  Measure the distance between your extruder and your mark and use that as your d value.
}
\examples{
resteps(82, 30)
}

