% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xysteps.R
\name{xysteps}
\alias{xysteps}
\title{printtools3d}
\usage{
xysteps(s, m, p, t)
}
\arguments{
\item{s}{The number of steps per revolution of your motor.}

\item{m}{Your driver's microstep setting.}

\item{p}{The angle of your belt pitch in steps/revolution (can be found with 360 / step angle in degrees).}

\item{t}{The number of teeth on your pulley.}
}
\description{
Calculate proper steps for X and Y axes in steps/mm.
}
\examples{
xysteps(200, 16, 2, 20)
}

