% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyztogrid.R
\name{xyztogrid}
\alias{xyztogrid}
\title{printtools3d}
\arguments{
\item{x}{The X values}

\item{y}{The Y values}

\item{z}{The Z values}
}
\description{
Converts XYZ columns to a grid matrix
}
\examples{
xyztogrid()
}

