% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsteps.R
\name{zsteps}
\alias{zsteps}
\title{zsteps}
\usage{
zsteps(s, m, t)
}
\arguments{
\item{s}{The number of steps per revolution of your motor.}

\item{m}{Your driver's microstep setting.}

\item{t}{The pitch of your lead screw.}
}
\description{
Calculate Z steps for leadscrew systems.
}
\examples{
zsteps(200, 16, .8)
}

