% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blgcode.R, R/graphBed.R
\name{blgcode}
\alias{blgcode}
\title{printtools3d}
\usage{
blgcode(x, y)
}
\arguments{
\item{x}{The size of your X axis.}

\item{y}{The size of your Y axis.}
}
\description{
Generate GCODE that will test the overall flatness of your print bed.

Generate GCODE that will test the overall flatness of your print bed.
}
\examples{
blgcode(150,150)
graphBed()
}

